ceph: perform asynchronous unlink if we have sufficient caps
When performing an unlink, if we have Fx on the parent directory and
Lx on the inode of the dentry being unlinked then we don't need to
wait on the reply before returning.

In that situation, just fix up the dcache and link count and return
immediately after issuing the call to the MDS. This does mean that we
need to hold an extra reference to the inode being unlinked, and extra
references to the caps to avoid races. Put those references in the
r_callback routine.

If the operation ends up failing, then set a writeback error on the
directory inode that can be fetched later by an fsync on the dir.

Signed-off-by: Jeff Layton <jlayton@kernel.org>