ceph: only send metrics when the MDS rank is ready
When the MDS rank is in clientreplay state these metrics requests
will be discard directly. Especially when there are a lot of known
client requests need to replay these useless metrics requests will
slow down the MDS rank.

With this we will send the metrics requests only when the MDS rank
is in active state.

URL: https://tracker.ceph.com/issues/61524
Signed-off-by: Xiubo Li <xiubli@redhat.com>